Adds deps for PostGIS
 * JSON-C
 * LibXML2
 * GEOS

Note that GEOS has binding options for SWIG, Python, Ruby, and PHP, none
of which I built because it doesn't seem to have any options to keep
them contained. If this turns out to be a problem, I might have to
change this or at least look into it more.