Some suggested changes in the build script
Apparently building with MPI enabled but testing with it turned off. Depending on mpi implementation, this would run the tests with an MPI aware binary but for one process only without an mpiexec/mpirun command. The timedep examples in the regression test would not be supported by such a binary, as this part of the code is serial only at the moment.