Alters module_maker_v2 to genericise gcc-libs
In future we want modules to generally not depend on a specific gcc-libs version, as we transition up away from 4.9.2. This change makes the automatic module prerequisite determination in make_module_v2 strip off the version number from any "gcc-libs" module currently loaded when it's converted into a prereq statement.